==> source_code/00_woman_webscraping.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.by import By
url = 'http://itempage3.auction.co.kr/DetailView.aspx?ItemNo=B732029517&frm3=V2'
driver = webdriver.Chrome('../driver/chromedriver')
driver.get(url)

Review_button = driver.find_element(By.CSS_SELECTOR, '#tap_moving_2')
Review_button.click()
Review_latest_button = driver.find_element(By.CSS_SELECTOR, '#sectionVipReview > div > div.box__list-review.box__list-review--active > div.box__review-sort > ul > li:nth-child(2)')
Review_latest_button.click()

import time

Reviews = []
for page in range(2,7):
    Review_body = driver.find_element(By.CSS_SELECTOR, '#divVipReview > ul')
    Review_list = Review_body.find_elements(By.CSS_SELECTOR, 'li > div > div.box__content > div.box__review-text > p')
    Review_button.click()

    for idx, review in enumerate(Review_list):
        try:
            Reviews.append(review.text)
        except:
            print(review)
    time.sleep(2)

    Review_page_button = driver.find_element(By.CSS_SELECTOR, f'#divVipReview > div > a:nth-child({page})')
    Review_page_button.click()
    time.sleep(2)


2페이지부터 7페이지까지는 다 숫자로 되어있지만 다음 페이지로 넘어가면 숫자가 고정된다.

for page in range(8,250):
    Review_page_button = driver.find_element(By.CSS_SELECTOR, f'#divVipReview > div > a:nth-child(5)')
    Review_page_button.click()
    time.sleep(2)
    Review_body = driver.find_element(By.CSS_SELECTOR, '#divVipReview > ul')
    Review_list = Review_body.find_elements(By.CSS_SELECTOR, 'li > div > div.box__content > div.box__review-text > p')
    Review_button.click()

    for idx, review in enumerate(Review_list):
        try:
            Reviews.append(review.text)
        except:
            print(review)
    time.sleep(2)



Reviews

len(Reviews)

Reviews[0]

Reviews[:5]

for idx, review in enumerate(Reviews):
    Reviews[idx] = review.replace("\n", " ")

Reviews[:5]

import pandas as pd

Revies_df = pd.DataFrame(Reviews)

Revies_df.head()

Revies_df.to_csv('woman.csv')
