==> source_code/01_practice_semantic_analysis.ipynb <==
from sklearn.feature_extraction.text import CountVectorizer
from konlpy.tag import Okt
import re
import pandas as pd
import numpy as np

df = pd.read_csv("../../data/naver_shopping.txt", sep = '\t', encoding = 'utf-8', header = None)

stopwords = pd.read_csv("https://raw.githubusercontent.com/yoonkt200/FastCampusDataset/master/korean_stopwords.txt").values.tolist()
stopwords.append('배송')

df.isnull().sum()

df.columns = ['rating', 'text']

len(df['text'])

def rating_to_label(rating):
    if rating >=4 :
        return 1
    else:
        return 0

df['y'] = df['rating'].apply(lambda x: rating_to_label(x))

tagger = Okt()
line = tagger.pos(df['text'][1])
line 

def text_preprocessed(text):
    hangul = re.compile('[^ ㄱ-ㅣ가-힣]')
    result = hangul.sub('', text)
    okt = Okt()
    Okt_morphs = okt.pos(result)
    words = []

    for word, pos in Okt_morphs:
        if pos == 'Adjective' or pos == 'Verb' or pos == 'Noun':
            if len(word) > 1 and word not in stopwords:
                words.append(word)

    words_str = ' '.join(words)
    return words_str

X_texts = []

for idx, row in df.iterrows():
    tokenized_review = text_preprocessed(df.loc[idx, 'text'])
    X_texts.append(tokenized_review)

df['tokenized'] = X_texts

list = df['tokenized'][2]
words_str = ' '.join(list)
words_str  

from tqdm import tqdm
import time

list_to_string = []
for idx, row in tqdm(df.iterrows()):
    list_to_string.append(' '.join(df.loc[idx, 'tokenized']))

list_to_string[:5]

df['tokenized'] = list_to_string
df.head()

df.to_csv('../../data/naver_shopping_tokenized.csv', sep = ",", encoding = 'utf-8')

## DTM 적용

from sklearn.model_selection import train_test_split
y = df['y'].values.tolist()
X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['tokenized'], y, test_size = 0.2, random_state = 0)

def tokenizer(text):
    return text.split()

vect = CountVectorizer(min_df = 3, ngram_range=(1, 2))
X_train_tf = vect.fit_transform(X_train_texts)
X_test_tf = vect.transform(X_test_texts)

vocablist = [word for word, number in sorted(vect.vocabulary_.items(), key=lambda x:x[1])]  

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from sklearn.model_selection import GridSearchCV

lr = LogisticRegression(random_state = 0)
lr.fit(X_train_tf, y_train)

y_pred = lr.predict(X_test_tf)

print("accuracy: %.2f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

## TFIDF 적용

y = df['y']
X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['tokenized'], y, test_size = 0.2, random_state = 0)

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf_vectorizer = TfidfVectorizer(ngram_range=(1,2), min_df = 3, max_df = 0.9)
tfidf_vectorizer.fit(X_train_texts)
tfidf_matrix_train = tfidf_vectorizer.transform(X_train_texts)


lr_tfidf = LogisticRegression(random_state = 0)
lr_tfidf.fit(tfidf_matrix_train, y_train)

tfidf_matrix_test = tfidf_vectorizer.transform(X_test_texts)
y_pred = lr_tfidf.predict(tfidf_matrix_test)

print("accuracy: %.2f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

## DTM 기준으로 하이퍼파라미터 튜닝 전 성능 평가
* accuracy: 0.78
* Precision : 0.742
* Recall : 0.850
* F1 : 0.792

## TFIDF 기준으로 하이퍼파라미터 튜닝 전 성능 평가
* accuracy: 0.78
* Precision : 0.750
* Recall : 0.835
* F1 : 0.790
