==> source_code/05_model_save_and_load.ipynb <==
from sklearn.feature_extraction.text import CountVectorizer
from konlpy.tag import Okt
import pandas as pd
import pickle

df = pd.read_csv("../../data/naver_shopping_tokenized.csv", encoding = 'utf-8', index_col = 0)

df.head()

from sklearn.model_selection import train_test_split
y = df['y'].values.tolist()
X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['text'], y, test_size = 0.2, random_state = 0)

okt = Okt()
def okt_tokenizer(text):
    tokens_ko = okt.morphs(text)
    return tokens_ko

## DTM

vect = CountVectorizer(tokenizer = okt_tokenizer, min_df = 3, ngram_range=(1,2))
X_train_tf = vect.fit_transform(X_train_texts)
X_test_tf = vect.transform(X_test_texts)

with open('saved_CounterVectorizer.pickle','wb') as fw:
    pickle.dump(vect, fw)

from sklearn.linear_model import LogisticRegression
lr_DTM = LogisticRegression(random_state = 0, solver = 'liblinear', C = 1)
lr_DTM.fit(X_train_tf, y_train)
y_pred_DTM = lr_DTM.predict(X_test_tf)

with open('saved_LogisticRegression_DTM.pickle','wb') as fw:
    pickle.dump(lr_DTM, fw)

## TFIDF

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf_vectorizer = TfidfVectorizer(ngram_range=(1,2), min_df = 3, max_df = 0.9, tokenizer = okt_tokenizer)
tfidf_vectorizer.fit(X_train_texts)
tfidf_matrix_train = tfidf_vectorizer.transform(X_train_texts)
tfidf_matrix_test = tfidf_vectorizer.transform(X_test_texts)

with open('saved_TfidfVectorizer.pickle','wb') as fw:
    pickle.dump(tfidf_vectorizer, fw)

lr_tfidf = LogisticRegression(random_state = 0, solver = 'liblinear', C = 3.5)
lr_tfidf.fit(tfidf_matrix_train, y_train)

with open('saved_LogisticRegression_TFIDF.pickle','wb') as fw:
    pickle.dump(lr_tfidf, fw)

## 크롤링한 남성용 의류 데이터 감성분류(DTM)

man = pd.read_csv("./man.csv", encoding = 'utf-8', index_col = 0)

man.info()

man.head()

with open('saved_CounterVectorizer.pickle','rb') as f:
    countvectorizer = pickle.load(f)

vectorized_review_m = countvectorizer.transform(man[man.columns[0]])

with open('saved_LogisticRegression_DTM.pickle','rb') as f:
    lr_DTM = pickle.load(f)

semantic_analysis_m = lr_DTM.predict(vectorized_review_m)

man['semantic'] = semantic_analysis_m

man['semantic'].value_counts()

man.columns = ['text', 'semantic_DTM']

## 크롤링한 여성의류 데이터 감성분류(DTM)

woman = pd.read_csv("./woman.csv", encoding = 'utf-8', index_col =  0)

woman.info()

woman.head()

vectorized_review_w = countvectorizer.transform(woman[woman.columns[0]])

semantic_analysis_w = lr_DTM.predict(vectorized_review_w)

woman['semantic'] = semantic_analysis_w

woman['semantic'].value_counts()

woman.columns = ['text', 'semantic_DTM']

## 크롤링한 남성용 의류 데이터 감성분류(TFIDF)

with open('saved_TfidfVectorizer.pickle','rb') as f:
    Tfidfvectorizer = pickle.load(f)

vectorized_review_m_2 = Tfidfvectorizer.transform(man['text'])

with open('saved_LogisticRegression_TFIDF.pickle', 'rb') as f:
    lr_tfidf = pickle.load(f)

predict = lr_tfidf.predict(vectorized_review_m_2)

man['semantic_TFIDF'] = predict

man['semantic_TFIDF'].value_counts()

## 크롤링한 여성의류 데이터 감성분류(TFIDF)


vectorized_review_w_2 = Tfidfvectorizer.transform(woman['text'])

predict2 = lr_tfidf.predict(vectorized_review_w_2)

woman['semantic_TFIDF'] = predict2

woman['semantic_TFIDF'].value_counts()

## TFIDF와 DTM 비교

df_merge = pd.concat([man, woman])

df_merge.head()

df_merge.reset_index(inplace= True, drop = True)

df_merge.groupby(['semantic_DTM', 'semantic_TFIDF']).count()

df_merge.to_csv("man_and_woman_predict_semantic_analysis.csv", encoding= 'utf-8')
