==> source_code/06_improved_preprocessing.ipynb <==
import pandas as pd

df = pd.read_csv("../../data/naver_shopping.csv", encoding = 'utf-8', index_col = 0)

df.head()

df.info()

wordList = ['좋은데요', '좋습니다', '조아요','좋네요', '좋고','졸습니다', '나쁘지 않네요', '조아','좋다']
change = '좋아요'

for word in wordList:
    df['Review'] = df['Review'].str.replace(word, change)

df.loc[500:520, 'Review']

wordList = ['까끌','까실까실','까슬까슬','거칠다','깔깔','깔','꺼끌거리다','가칠','까슬하다','까칠','까칠까칠','꺼끌','꺼끌꺼끌','꺼칠','껄끄럽다','거슬거리다','가슬','가실','따끔']
change = '따갑다'

for word in wordList:
    df['Review'] = df['Review'].str.replace(word, change)

wordList = ['맛있습니다','맛있네요','맛있고']
change = '맛있어요'

for word in wordList:
    df['Review'] = df['Review'].str.replace(word, change)

f = open("../../data/review_synonym.txt", 'r', encoding= 'utf-8')

lines = f.readlines()

for idx, l in enumerate(lines):
    lines[idx] = l.strip()

lines[0].split(',')

lines_dic = {}

for idx, l in enumerate(lines):
    temp = l.split(',')
    lines_dic[temp[0]] = temp[1:]

for key, wordList in lines_dic.items():
    for word in wordList:
        df['Review'] = df['Review'].str.replace(word, key)

df.to_csv("../../data/naver_shopping_preprocessed.csv")

df_man = pd.read_csv("../../data/man.csv", encoding = 'utf-8')

df_man.head()

df_man[df_man.columns[1]]
