==> source_code/08_tokenized_DF_production.ipynb <==
import pandas as pd

df = pd.read_csv("../../data/naver_shopping_tokenized_review.csv")

df.info()

df = df[df['tokenized_1'].notnull() & df['tokenized_2'].notnull()]

### DTM prediction column 추가(tokenizer1)
* 예측치 of (DTM + 명사, 형용사, 동사만 추출 & 2글자 이상만 토큰화)

from sklearn.model_selection import train_test_split
y = df['y']
X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['tokenized_1'], y, test_size = 0.2, random_state = 0)

def tokenizer(text):
    return text.split()

from sklearn.feature_extraction.text import CountVectorizer

vect = CountVectorizer(min_df = 3, ngram_range=(1,2), tokenizer = tokenizer)
X_train_tf = vect.fit_transform(X_train_texts)
X_total = vect.transform(df['tokenized_1'])

import pickle
with open('DTM_tokenizer1_gridsearch_done.pickle','rb') as f:
    lr_DTM_tk1 = pickle.load(f)

y_pred = lr_DTM_tk1.predict(X_total)

# del df[df.columns[0]]
df.head()

df['DTM_tk1_prediction'] = y_pred

### TFIDF prediction column 추가(tokenizer1)
* 예측치 of (TFIDF + 명사, 형용사, 동사만 추출 & 2글자 이상만 토큰화)

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf_vectorizer = TfidfVectorizer(ngram_range=(1,2), min_df = 3, max_df = 0.9, tokenizer = tokenizer)
tfidf_vectorizer.fit(X_train_texts)
tfidf_matrix_X_total = tfidf_vectorizer.transform(df['tokenized_1'])

with open('TFIDF_tokenizer1_gridsearch_done.pickle','rb') as f:
    lr_TFIDF_tk1 = pickle.load(f)

y_pred2 = lr_TFIDF_tk1.predict(tfidf_matrix_X_total)

df['TFIDF_tk1_prediction'] = y_pred2

df.head()

### DTM prediction column 추가(tokenizer2)
* 예측치 of (DTM + 형태소 모두 활용)

X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['tokenized_2'], y, test_size = 0.2, random_state = 0)

vect2 = CountVectorizer(min_df = 3, ngram_range=(1,2), tokenizer = tokenizer)
X_train_tf = vect2.fit_transform(X_train_texts)
X_total_tf = vect2.transform(df['tokenized_2'])

with open('DTM_tokenizer2_gridsearch_done.pickle','rb') as f:
    lr_DTM_tk2 = pickle.load(f)

y_pred3 = lr_DTM_tk2.predict(X_total_tf)

df['DTM_tk2_prediction'] = y_pred3

df.head()

### TFIDF prediction column 추가(tokenizer2)
* 예측치 of (TFIDF + 형태소 모두 활용)

tfidf_vectorizer2 = TfidfVectorizer(ngram_range=(1,2), min_df = 3, max_df = 0.9, tokenizer = tokenizer)
tfidf_vectorizer2.fit(X_train_texts)
tfidf_matrix_X_total = tfidf_vectorizer2.transform(df['tokenized_2'])

with open('TFIDF_tokenizer2_gridsearch_done.pickle','rb') as f:
    lr_TFIDF_tk2 = pickle.load(f)

y_pred4 = lr_TFIDF_tk2.predict(tfidf_matrix_X_total)

df['TFIDF_tk2_prediction'] = y_pred4

df.info()

df.to_csv("../../data/naver_shopping_tokenized_review_and_prediction_by_models.csv", sep = ",", encoding ="utf-8")
