==> source_code/09_Modeling_final.ipynb <==
from sklearn.feature_extraction.text import CountVectorizer
from konlpy.tag import Okt
import pandas as pd

## tokenized된 리뷰 데이터 전처리 보완하기
* 불러오기
* null 값 있는 컬럼 제거하기
* 유의어 일원화 보완
* 갱신된 불용어 사전 적용

df = pd.read_csv("../../data/naver_shopping_tokenized_review.csv", encoding ='utf-8', index_col = 0)
df[100:120]

df.isnull().sum()

df = df[df['tokenized_1'].notnull() & df['tokenized_2'].notnull()]
df.isnull().sum()

wordList = ['만족합니다','만족스럽습니다','만족스럽네요','만족합니다','만족스러워요']
change = '만족해요'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

wordList = ['좋구요','좋네요','좋습니다','좋아해요','좋아합니다','좋아요요']
change = '좋아요'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

wordList = ['안좋아요요']
change = '안좋아요'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

wordList = ['에드', '맘에듭니']
change = '마음에 듭니다'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

wordList = ['좋아여', '좋아용','좋아하시네요','좋아요하시네요','좋아요요']
change = '좋아요'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

wordList = ['맛있었어요']
change = '맛있어요'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

wordList = ['예쁘다','이쁘고','예쁘고']
change = '예뻐요'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

wordList = ['좋았어요','좋으네요']
change = '좋아요'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

wordList = ['느리네요','느리고','느려서']
change = '느려요'

for word in wordList:
    df['tokenized_1'] = df['tokenized_1'].str.replace(word, change, regex = True)
    df['tokenized_2'] = df['tokenized_2'].str.replace(word, change, regex = True)

# stopwords = pd.read_csv("https://raw.githubusercontent.com/yoonkt200/FastCampusDataset/master/korean_stopwords.txt").values.tolist()
# temp = []
# for s in stopwords:
#     temp.append(s[0])

f = open("../../data/stopwords_self_improved.txt", 'r', encoding= 'utf-8')

lines = f.readlines()
for idx, l in enumerate(lines):
    lines[idx] = l.strip()

stopwords = []
for l in lines:
    stopwords.append(l)



import time
from tqdm import tqdm

for stopword in tqdm(stopwords):
    df.replace(stopword, '', regex = True, inplace = True)

df.to_csv("../../data/naver_shopping_tokenized_review_improved.csv", sep = ",", encoding = 'utf-8')

## 토큰화 함수 정의
* 원래는 CounterVectorizer나 TFIDF 하이퍼파라미터에 해당 tokenizer_1, tokenizer_2 함수를 tokenizer로 지정하는 방법이 기본적인 방법
* 시간을 효율적으로 쓰기 위해 tokenizer_1, tokenizer_2 함수를 적용한 컬럼 tokenized_1, tokenized_2를 df에 추가함
* 명사, 형용사, 동사 + 1글자 이상만 추출하는 경우 - tokenized_1 리뷰 데이터 활용
* 형태소 모두 사용 - tokenized_2 리뷰 데이터 활용
* vectorized 클래스에 들어갈 tokenizer는 띄어쓰기 기준으로 split하는 기능만 적용

def tokenizer(text):
    return text.split()

### 토큰화된 데이터 프레임 불러오기

df[250:270]

print(df.isnull().sum())
df.info()

## DTM + GridSearchCV + tokenizer_1
* countervectorizer
* GridSearchCV
* Tokenizer1

from sklearn.model_selection import train_test_split
y = df['y']
X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['tokenized_1'], y, test_size = 0.2, random_state = 0)

vect = CountVectorizer(min_df = 3, ngram_range=(1,2), tokenizer = tokenizer)
X_train_tf = vect.fit_transform(X_train_texts)
X_test_tf = vect.transform(X_test_texts)

vocablist_DTM_tk1 = [word for word, number in sorted(vect.vocabulary_.items(), key = lambda x:x[1])]

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from sklearn.model_selection import GridSearchCV

lr = LogisticRegression(random_state = 0, solver = 'liblinear', max_iter = 1000)
params = { 'C': [0.6, 0.8, 1 ,3.5, 4.5 ] }
grid_cv = GridSearchCV(lr , param_grid=params , cv=3 ,scoring='accuracy', verbose=1 )
grid_cv.fit(X_train_tf , y_train)
print(grid_cv.best_params_ , round(grid_cv.best_score_,4))

best_estimator_DTM_tk1 = grid_cv.best_estimator_
y_pred = best_estimator_DTM_tk1.predict(X_test_tf)

print("accuracy: %.2f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

import pickle

with open('DTM_tokenizer1_gridsearch_done_improved.pickle','wb') as fw:
    pickle.dump(best_estimator_DTM_tk1, fw)

coefficients = best_estimator_DTM_tk1.coef_.tolist()

sorted_coefficients = sorted(enumerate(coefficients[0]), key=lambda x:x[1], reverse=True)
# coefficients(계수)가 큰 값부터 내림차순으로 정렬

print('긍정적인 단어 Top 20 (높은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[:20]:
  print('{0:}({1:.3f})'.format(vocablist_DTM_tk1[word_num], coef))

print('\n부정적인 단어 Top 20 (낮은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[-20:][::-1]: 
  print('{0:}({1:.3f})'.format(vocablist_DTM_tk1[word_num], coef))

## TFIDF + GridSearchCV + tokenizer_1
* TFIDF
* GridSearchCV
* Tokenizer1

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf_vectorizer = TfidfVectorizer(ngram_range=(1,2), min_df = 3, max_df = 0.9, tokenizer = tokenizer)
tfidf_vectorizer.fit(X_train_texts)
tfidf_matrix_train = tfidf_vectorizer.transform(X_train_texts)
tfidf_matrix_test = tfidf_vectorizer.transform(X_test_texts)

vocablist_TFIDF_tk1 = [word for word, number in sorted(tfidf_vectorizer.vocabulary_.items(), key = lambda x:x[1])]

lr = LogisticRegression(random_state = 0, solver = 'liblinear', max_iter = 1000)
params = { 'C': [0.6, 0.8, 1 ,3.5, 4.5 ] }
grid_cv = GridSearchCV(lr , param_grid=params , cv=3 ,scoring='accuracy', verbose=1 )
grid_cv.fit(tfidf_matrix_train , y_train)
print(grid_cv.best_params_ , round(grid_cv.best_score_,4))

best_estimator_TFIDF_tk1 = grid_cv.best_estimator_
y_pred = best_estimator_TFIDF_tk1.predict(tfidf_matrix_test)

print("accuracy: %.2f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

with open('TFIDF_tokenizer1_gridsearch_done_improved.pickle','wb') as fw:
    pickle.dump(best_estimator_TFIDF_tk1, fw)

coefficients = best_estimator_TFIDF_tk1.coef_.tolist()

sorted_coefficients = sorted(enumerate(coefficients[0]), key=lambda x:x[1], reverse=True)
# coefficients(계수)가 큰 값부터 내림차순으로 정렬

print('긍정적인 단어 Top 20 (높은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[:20]:
  print('{0:}({1:.3f})'.format(vocablist_TFIDF_tk1[word_num], coef))

print('\n부정적인 단어 Top 20 (낮은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[-20:][::-1]: 
  print('{0:}({1:.3f})'.format(vocablist_TFIDF_tk1[word_num], coef))

## DTM + GridSearchCV + tokenizer_2
* countervectorizer
* GridSearchCV
* Tokenizer2

X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['tokenized_2'], y, test_size = 0.2, random_state = 0)

vect2 = CountVectorizer(min_df = 3, ngram_range=(1,2), tokenizer = tokenizer)
X_train_tf = vect2.fit_transform(X_train_texts)
X_test_tf = vect2.transform(X_test_texts)

vocablist_DTM_tk2 = [word for word, number in sorted(vect2.vocabulary_.items(), key = lambda x:x[1])]

lr = LogisticRegression(random_state = 0, solver = 'liblinear', max_iter = 1000)
params = { 'C': [0.6, 0.8, 1 ,3.5, 4.5 ] }
grid_cv = GridSearchCV(lr , param_grid=params , cv=3 ,scoring='accuracy', verbose=1 )
grid_cv.fit(X_train_tf , y_train)
print(grid_cv.best_params_ , round(grid_cv.best_score_,4))

best_estimator_DTM_tk2 = grid_cv.best_estimator_
y_pred = best_estimator_DTM_tk2.predict(X_test_tf)

print("accuracy: %.2f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

with open('DTM_tokenizer2_gridsearch_done_improved.pickle','wb') as fw:
    pickle.dump(best_estimator_DTM_tk2, fw)

coefficients = best_estimator_DTM_tk2.coef_.tolist()

sorted_coefficients = sorted(enumerate(coefficients[0]), key=lambda x:x[1], reverse=True)
# coefficients(계수)가 큰 값부터 내림차순으로 정렬

print('긍정적인 단어 Top 20 (높은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[:20]:
  print('{0:}({1:.3f})'.format(vocablist_DTM_tk2[word_num], coef))

print('\n부정적인 단어 Top 20 (낮은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[-20:][::-1]: 
  print('{0:}({1:.3f})'.format(vocablist_DTM_tk2[word_num], coef))

## TFIDF + GridSearchCV + tokenizer_2
* TFIDF
* GridSearchCV
* Tokenizer2

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf_vectorizer2 = TfidfVectorizer(ngram_range=(1,2), min_df = 3, max_df = 0.9, tokenizer = tokenizer)
tfidf_vectorizer2.fit(X_train_texts)
tfidf_matrix_train = tfidf_vectorizer2.transform(X_train_texts)
tfidf_matrix_test = tfidf_vectorizer2.transform(X_test_texts)

vocablist_TFIDF_tk2 = [word for word, number in sorted(vect2.vocabulary_.items(), key = lambda x:x[1])]

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from sklearn.model_selection import GridSearchCV

lr = LogisticRegression(random_state = 0, solver = 'liblinear', max_iter = 1000)
params = { 'C': [0.6, 0.8, 1 ,3.5, 4.5 ] }
grid_cv = GridSearchCV(lr , param_grid=params , cv=3 ,scoring='accuracy', verbose=1 )
grid_cv.fit(tfidf_matrix_train, y_train)
print(grid_cv.best_params_ , round(grid_cv.best_score_,4))

best_estimator_TFIDF_tk2 = grid_cv.best_estimator_
y_pred = best_estimator_TFIDF_tk2.predict(X_test_tf)
print("accuracy: %.2f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

import pickle

with open('TFIDF_tokenizer2_gridsearch_done_improved.pickle','wb') as fw:
    pickle.dump(best_estimator_TFIDF_tk2, fw)

coefficients = best_estimator_TFIDF_tk2.coef_.tolist()

sorted_coefficients = sorted(enumerate(coefficients[0]), key=lambda x:x[1], reverse=True)
# coefficients(계수)가 큰 값부터 내림차순으로 정렬

print('긍정적인 단어 Top 20 (높은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[:20]:
  print('{0:}({1:.3f})'.format(vocablist_TFIDF_tk2[word_num], coef))

print('\n부정적인 단어 Top 10 (낮은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[-20:][::-1]: 
  print('{0:}({1:.3f})'.format(vocablist_TFIDF_tk2[word_num], coef))

### DTM + 하이퍼파라미터 튜닝 + 유의어 일원화 보완(보완전) + 명사/형용사/동사만 추출
* accuracy: 0.90 (0.90)
* Precision : 0.905 (0.905)
* Recall : 0.891 (0.892)
* F1 : 0.898 (0.898)

### TFIDF + 하이퍼파라미터 튜닝 + 유의어 일원화 보완(보완전)+ 명사/형용사/동사만 추출
* accuracy: 0.90(0.90)
* Precision : 0.905(0.907)
* Recall : 0.893(0.894)
* F1 : 0.899(0.900)

### DTM + 하이퍼파라미터 튜닝 + 유의어 일원화 보완(보완전)+ 형태소만 추출
* accuracy: 0.91(0.91)
* Precision : 0.915(0.914)
* Recall : 0.910(0.910)
* F1 : 0.912 (0.912)

### TFIDF + 하이퍼파라미터 튜닝 + 유의어 일원화 보완(보완전)+ 형태소만 추출
* accuracy: 0.90(0.90)
* Precision : 0.884(0.883)
* Recall : 0.931(0.931)
* F1 : 0.907(0.906)
