==> source_code/12_From_Preprocessing_To_Modeling_Final_after_synonym_unification.ipynb <==
### 1. 데이터 전처리
#### 1-1. 데이터 load 
- https://github.com/bab2min/corpus/tree/master/sentiment 에서 다운로드한 데이터 다운로드 후 load
- columns 이름 추가
- label column 생성


import pandas as pd

df = pd.read_csv("../../data/naver_shopping.txt", sep = '\t', encoding = 'utf-8', header = None)

df.head()

df.isnull().sum()

df.columns = ['rating', 'Review']

df.head()

def rating_to_label(rating):
    if rating >=4 :
        return 1
    else:
        return 0

df['y'] = df['rating'].apply(lambda x: rating_to_label(x))

#### 1-2. 유의어 일원화
- 자체적으로 만든 유의어 사전 로딩(txt 파일)
- 비슷한 단어들을 하나의 단어로 통일하는 전처리 적용

f = open("../../data/review_synonym_final.txt", 'r', encoding= 'utf-8')

lines = f.readlines()

for idx, l in enumerate(lines):
    lines[idx] = l.strip()

lines_dic = {}

for idx, l in enumerate(lines):
    temp = l.split(',')
    lines_dic[temp[0]] = temp[1:]

for key, wordList in lines_dic.items():
    for word in wordList:
        df['Review'] = df['Review'].str.replace(word, key)

#### 1-3. 불용어 사전 편집
- "https://raw.githubusercontent.com/yoonkt200/FastCampusDataset/master/korean_stopwords.txt" 활용
- EDA 과정에서 평점과 상관없이 빈출된 단어(배송, 구매, 생각, 사용 등) 추가

stopwords = pd.read_csv("https://raw.githubusercontent.com/yoonkt200/FastCampusDataset/master/korean_stopwords.txt").values.tolist()
stopwords.append('배송')
stopwords.append('구매')
stopwords.append('생각')
stopwords.append('사용')
stopwords.append('가격')
stopwords.append('제품')

#### 1-4. 토큰화된 리뷰 컬럼 추가
- 명사, 동사, 형용사에 해당하는 2글자 이상의 형태소만 활용하는 토큰화함수 1을 적용한 리뷰 컬럼 생성
- 모든 형태소와 이모티콘을 활용하는 토큰화함수 2를 적용한 리뷰 컬럼 생성
- 향후 countervectorization, TFIDF vectorization에 사용하는 토큰화함수는 띄어쓰기 기준으로 토큰화하는 간소화된 함수 사용

from konlpy.tag import Okt

def tokenizer_1(text):
    hangul = re.compile('[^ ㄱ-ㅣ가-힣]')
    result = hangul.sub('', text)
    okt = Okt()
    Okt_morphs = okt.pos(result)
    words = []

    for word, pos in Okt_morphs:
        if pos == 'Adjective' or pos == 'Verb' or pos == 'Noun':
            if len(word) > 1 and word not in stopwords:
                words.append(word)

    return ' '.join(words)

okt2 = Okt()

def tokenizer_2(text):
    tokens_ko = okt2.morphs(text)

    result = []
    for word in tokens_ko:
        if word not in stopwords:
            result.append(word)
    return ' '.join(result)

import time
from tqdm import tqdm
import re

tokenized_1 = []

for idx, row in tqdm(df.iterrows()):
    tokenized_review = tokenizer_1(df.loc[idx, 'Review'])
    tokenized_1.append(tokenized_review)

df['tokenized_1'] = tokenized_1

tokenized_2 = []

for idx, row in tqdm(df.iterrows()):
    tokenized_review = tokenizer_2(df.loc[idx, 'Review'])
    tokenized_2.append(tokenized_review)

df['tokenized_2'] = tokenized_2

df.head()

df.to_csv("../../data/naver_shopping_preprocessed_final.csv", sep = ",", encoding = "utf-8")

print(df.isnull().sum())
df.info()

def tokenizer(text):
    return text.split()

### 2. 모델링
#### 2-1. DTM + GridSearchCV + tokenizer_1
* countervectorizer
* GridSearchCV
* Tokenizer1

from sklearn.model_selection import train_test_split
y = df['y']
X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['tokenized_1'], y, test_size = 0.2, random_state = 0)

from sklearn.feature_extraction.text import CountVectorizer

vect1 = CountVectorizer(min_df = 3, ngram_range=(1,2), tokenizer = tokenizer)
X_train_vectorized1 = vect1.fit_transform(X_train_texts)
X_test_vectorized1 = vect1.transform(X_test_texts)

vocablist_DTM_tk1 = [word for word, number in sorted(vect1.vocabulary_.items(), key = lambda x:x[1])]

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from sklearn.model_selection import GridSearchCV

lr = LogisticRegression(random_state = 0, solver = 'liblinear', max_iter = 1000)
params = { 'C': [0.6, 0.8, 1 , 3.5, 4.5 ] }
grid_cv = GridSearchCV(lr , param_grid=params , cv=3 ,scoring='accuracy', verbose=2 )
grid_cv.fit(X_train_vectorized1 , y_train)
print(grid_cv.best_params_ , round(grid_cv.best_score_,3))

best_estimator_DTM_tk1 = grid_cv.best_estimator_
y_pred = best_estimator_DTM_tk1.predict(X_test_vectorized1)

print("accuracy: %.3f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

import pickle

with open('saved_CounterVectorizer_tk1_final.pickle','wb') as fw:
    pickle.dump(vect1, fw)

with open('DTM_tokenizer1_gridsearch_done_final.pickle','wb') as fw:
    pickle.dump(best_estimator_DTM_tk1, fw)

coefficients = best_estimator_DTM_tk1.coef_.tolist()

sorted_coefficients = sorted(enumerate(coefficients[0]), key=lambda x:x[1], reverse=True)
# coefficients(계수)가 큰 값부터 내림차순으로 정렬

print('긍정적인 단어 Top 20 (높은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[:20]:
  print('{0:}({1:.3f})'.format(vocablist_DTM_tk1[word_num], coef))

print('\n부정적인 단어 Top 20 (낮은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[-20:][::-1]: 
  print('{0:}({1:.3f})'.format(vocablist_DTM_tk1[word_num], coef))

#### 2-2. TFIDF + GridSearchCV + tokenizer_1
* TFIDF
* GridSearchCV
* Tokenizer1

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf_vectorizer1 = TfidfVectorizer(ngram_range=(1,2), min_df = 3, max_df = 0.9, tokenizer = tokenizer)
tfidf_vectorizer1.fit(X_train_texts)
tfidf_matrix_train = tfidf_vectorizer1.transform(X_train_texts)
tfidf_matrix_test = tfidf_vectorizer1.transform(X_test_texts)

vocablist_TFIDF_tk1 = [word for word, number in sorted(tfidf_vectorizer1.vocabulary_.items(), key = lambda x:x[1])]

lr = LogisticRegression(random_state = 0, solver = 'liblinear', max_iter = 1000)
params = { 'C': [0.6, 0.8, 1 ,3.5, 4.5 ] }
grid_cv = GridSearchCV(lr , param_grid=params , cv=3 ,scoring='accuracy', verbose=2 )
grid_cv.fit(tfidf_matrix_train , y_train)
print(grid_cv.best_params_ , round(grid_cv.best_score_,3))

best_estimator_TFIDF_tk1 = grid_cv.best_estimator_
y_pred = best_estimator_TFIDF_tk1.predict(tfidf_matrix_test)

print("accuracy: %.3f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

with open('saved_TfidfVectorizer_tk1_final.pickle','wb') as fw:
    pickle.dump(tfidf_vectorizer1, fw)

with open('TFIDF_tokenizer1_gridsearch_done_final.pickle','wb') as fw:
    pickle.dump(best_estimator_TFIDF_tk1, fw)

coefficients = best_estimator_TFIDF_tk1.coef_.tolist()

sorted_coefficients = sorted(enumerate(coefficients[0]), key=lambda x:x[1], reverse=True)
# coefficients(계수)가 큰 값부터 내림차순으로 정렬

print('긍정적인 단어 Top 20 (높은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[:20]:
  print('{0:}({1:.3f})'.format(vocablist_TFIDF_tk1[word_num], coef))

print('\n부정적인 단어 Top 20 (낮은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[-20:][::-1]: 
  print('{0:}({1:.3f})'.format(vocablist_TFIDF_tk1[word_num], coef))

#### 2-3. DTM + GridSearchCV + tokenizer_2
* countervectorizer
* GridSearchCV
* Tokenizer2

X_train_texts, X_test_texts, y_train, y_test = train_test_split(df['tokenized_2'], y, test_size = 0.2, random_state = 0)

vect2 = CountVectorizer(min_df = 3, ngram_range=(1,2), tokenizer = tokenizer)
X_train_vectorized2 = vect2.fit_transform(X_train_texts)
X_test_vectorized2 = vect2.transform(X_test_texts)

vocablist_DTM_tk2 = [word for word, number in sorted(vect2.vocabulary_.items(), key = lambda x:x[1])]

lr = LogisticRegression(random_state = 0, solver = 'liblinear', max_iter = 1000)
params = { 'C': [0.6, 0.8, 1 ,3.5, 4.5 ] }
grid_cv = GridSearchCV(lr , param_grid=params , cv=3 ,scoring='accuracy', verbose=2 )
grid_cv.fit(X_train_vectorized2 , y_train)
print(grid_cv.best_params_ , round(grid_cv.best_score_,3))

best_estimator_DTM_tk2 = grid_cv.best_estimator_
y_pred = best_estimator_DTM_tk2.predict(X_test_vectorized2)

print("accuracy: %.3f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

with open('saved_CounterVectorizer_tk2_final.pickle','wb') as fw:
    pickle.dump(vect2, fw)

with open('DTM_tokenizer2_gridsearch_done_final.pickle','wb') as fw:
    pickle.dump(best_estimator_DTM_tk2, fw)

coefficients = best_estimator_DTM_tk2.coef_.tolist()

sorted_coefficients = sorted(enumerate(coefficients[0]), key=lambda x:x[1], reverse=True)
# coefficients(계수)가 큰 값부터 내림차순으로 정렬

print('긍정적인 단어 Top 20 (높은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[:20]:
  print('{0:}({1:.3f})'.format(vocablist_DTM_tk2[word_num], coef))

print('\n부정적인 단어 Top 20 (낮은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[-20:][::-1]: 
  print('{0:}({1:.3f})'.format(vocablist_DTM_tk2[word_num], coef))

#### 2-4. TFIDF + GridSearchCV + tokenizer_2
* TFIDF
* GridSearchCV
* Tokenizer2

tfidf_vectorizer2 = TfidfVectorizer(ngram_range=(1,2), min_df = 3, max_df = 0.9, tokenizer = tokenizer)
tfidf_vectorizer2.fit(X_train_texts)
tfidf_matrix_train = tfidf_vectorizer2.transform(X_train_texts)
tfidf_matrix_test = tfidf_vectorizer2.transform(X_test_texts)

vocablist_TFIDF_tk2 = [word for word, number in sorted(vect2.vocabulary_.items(), key = lambda x:x[1])]

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
from sklearn.model_selection import GridSearchCV

lr = LogisticRegression(random_state = 0, solver = 'liblinear', max_iter = 1000)
params = { 'C': [0.6, 0.8, 1 ,3.5, 4.5 ] }
grid_cv = GridSearchCV(lr , param_grid=params , cv=3 ,scoring='accuracy', verbose=1 )
grid_cv.fit(tfidf_matrix_train, y_train)
print(grid_cv.best_params_ , round(grid_cv.best_score_,3))

best_estimator_TFIDF_tk2 = grid_cv.best_estimator_
y_pred = best_estimator_TFIDF_tk2.predict(tfidf_matrix_test)
print("accuracy: %.3f" % accuracy_score(y_test, y_pred))
print("Precision : %.3f" % precision_score(y_test, y_pred))
print("Recall : %.3f" % recall_score(y_test, y_pred))
print("F1 : %.3f" % f1_score(y_test, y_pred))

with open('saved_TfidfVectorizer_tk2_final.pickle','wb') as fw:
    pickle.dump(tfidf_vectorizer2, fw)

with open('TFIDF_tokenizer2_gridsearch_done_final.pickle','wb') as fw:
    pickle.dump(best_estimator_TFIDF_tk2, fw)

coefficients = best_estimator_TFIDF_tk2.coef_.tolist()

sorted_coefficients = sorted(enumerate(coefficients[0]), key=lambda x:x[1], reverse=True)
# coefficients(계수)가 큰 값부터 내림차순으로 정렬

print('긍정적인 단어 Top 20 (높은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[:20]:
  print('{0:}({1:.3f})'.format(vocablist_TFIDF_tk2[word_num], coef))

print('\n부정적인 단어 Top 10 (낮은 평점과 상관관계가 강한 단어들)')
for word_num, coef in sorted_coefficients[-20:][::-1]: 
  print('{0:}({1:.3f})'.format(vocablist_TFIDF_tk2[word_num], coef))
